final_pjt/final_pjt_back/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
oauthlib==3.2.2
openai==1.30.1
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
